backend/requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
boto3==1.40.36
botocore==1.40.36
cachetools==6.2.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.7
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
eng_to_ipa==0.0.2
fastapi==0.116.1
google-ai-generativelanguage==0.6.10
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
google-generativeai==0.8.3
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jmespath==1.0.1
jwt==1.4.0
Mako==1.3.10
MarkupSafe==3.0.2
oauthlib==3.3.1
packaging==25.0
passlib==1.7.4
payos==1.0.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
pytest==8.4.2
pytest-asyncio==1.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
s3transfer==0.14.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.11.0
SQLAlchemy==2.0.43
starlette==0.47.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
uuid-v7==1.0.0
uuid7==0.1.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
